adding steps and ingredients works

+++ app/src/main/java/p/l/omnomnom/igredient/Ingredient.java
@@ -1,7 +1,7 @@
 package p.l.omnomnom.igredient;
 
 public class Ingredient {
-    private int id;
+    private long id;
     private String name;
 
     public Ingredient(){
@@ -12,16 +12,16 @@ public class Ingredient {
         this.name = name;
     }
 
-    public Ingredient(int id, String name){
+    public Ingredient(long id, String name){
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
